test(threads): 拆分 useThreadsReducer 超长测试文件并保持行为不变
背景：\n-  超过 3000 行，不利于维护与定位\n- 本次按测试主题进行物理拆分，降低单文件复杂度并提升可读性\n\n改动说明：\n- 将线程列表 / pending 保留 / rename 合并 / tool 状态收敛 / plan 收敛相关用例迁移至新文件\n  \n- 原文件保留前段核心 reducer 测试，行数压降至 2500 以下\n- 仅做测试文件重组，不改生产代码与断言语义\n\n结果：\n- ：2435 行\n- 新增 ：828 行\n\n验证：\n- 执行
 RUN  v3.2.4 /Users/chenxiangning/code/AI/github/mossx

 ✓ src/features/threads/hooks/useThreadsReducer.threadlist-pending.test.ts (18 tests) 9ms
 ✓ src/features/threads/hooks/useThreadsReducer.test.ts (70 tests) 19ms

 Test Files  2 passed (2)
      Tests  88 passed (88)
   Start at  19:14:13
   Duration  443ms (transform 174ms, setup 75ms, collect 305ms, tests 28ms, environment 0ms, prepare 121ms)\n- 结果 88/88 通过，确认拆分前后行为一致
